requirements.txt
certifi==2018.4.16
chardet==3.0.4
crispy==0.5.0
Django==2.0.5
django-bootstrap-form==3.4
django-ckeditor==5.5.0
django-crispy-forms==1.7.2
django-embed-video==1.1.2
django-extensions==2.0.7
django-filter==1.1.0
django-js-asset==1.1.0
django-tinymce==2.7.0
django-wysiwyg==0.8.0
idna==2.6
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
